src/stack.py
from typing import Any


class Node:
    """Класс для узла стека"""

    def __init__(self, data, next_node):
        """
        Конструктор класса Node

        :param data: данные, которые будут храниться в узле
        """
        self.data: Any = data
        self.next_node: Node | None = next_node


class Stack:
    """Класс для стека"""

    def __init__(self):
        """Конструктор класса Stack"""
        self.top: Node | None = None

    def push(self, data):
        """
        Метод для добавления элемента на вершину стека

        :param data: данные, которые будут добавлены на вершину стека
        """
        self.top = Node(data, self.top)

    def pop(self):
        """
        Метод для удаления элемента с вершины стека и его возвращения

        :return: данные удаленного элемента
        """
        prev_top = self.top
        self.top = self.top.next_node
        return prev_top.data



# stack = Stack()
# print(stack.top)  # None
#
# stack.push('data1')
# print(stack.top.data)  # data1
#
# stack.push('data2')
# print(stack.top.data) # data2
#
# stack.push('data3')
# print(stack.top.data) # data3
#
